refactor: schedule components
